+++ keystar/src/Engine/Clock.h
@@ -3,12 +3,12 @@
 
 struct Clock
 {
-    uint32_t last_tick_time = 0;
+    Uint64 last_tick_time = 0;
     float delta = 0;
 
     void tick()
     {
-        uint32_t tick_time = SDL_GetTicks();
+        Uint64 tick_time = SDL_GetTicks();
         delta = (tick_time - last_tick_time) / 1000.0f;
         last_tick_time = tick_time;
     }
